Used conftest file to avoid importing pytest fixtures, and removed print statement.
