+++ improver_tests/developer_tools/test_MOMetadataInterpreter.py
@@ -28,36 +28,11 @@
 # CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
-"""Unit tests for the MOMetadataInterpreter plugin
-
-Description of imports: fixture name -> cube name:
-- ensemble_fixture -> ensemble_cube
-- landmask_fixture -> landmask_cube
-- percentile_fixture -> wind_gust_percentile_cube
-- probability_above_fixture -> probability_above_cube
-- probability_below_fixture -> blended_probability_below_cube
-- snow_level_fixture -> snow_level_cube
-- spot_fixture -> blended_spot_median_cube
-- wind_direction_fixture -> wind_direction_cube
-- wxcode_fixture -> wxcode_cube
-"""
+"""Unit tests for the MOMetadataInterpreter plugin"""
 
 import numpy as np
 import pytest
 from iris.coords import CellMethod
-
-from improver_tests.developer_tools import (
-    ensemble_fixture,
-    interpreter_fixture,
-    landmask_fixture,
-    percentile_fixture,
-    probability_above_fixture,
-    probability_below_fixture,
-    snow_level_fixture,
-    spot_fixture,
-    wind_direction_fixture,
-    wxcode_fixture,
-)
 
 ## Test successful outputs (input cubes in alphabetical order by fixture)
 
@@ -251,7 +226,6 @@ def test_warning_wind_gust_attribute_wrong_diagnostic(
     on an unexpected diagnostic"""
     wind_gust_percentile_cube.rename("wind_speed")
     interpreter.run(wind_gust_percentile_cube)
-    print(interpreter.warnings)
     assert interpreter.warnings == [
         "dict_keys(['source', 'title', 'institution', 'mosg__model_configuration', "
         "'wind_gust_diagnostic']) include unexpected attributes ['wind_gust_diagnostic']. "
